Updated preliminary analysis
Updated fxn getTaskGNG() to process data from a cell array of multiple sessions. We now have a block of sessions picked out for both Da and Le for the paper. These are the input to this fxn. Also updated labels on some fields in TEMPO_EV(). Will need to continue to do so with more reading into the TEMPO code.